Refactor frontend structure, renaming, fix backend bugs
