+++ client/src/components/content/BpmnModeler/BpmnModeler.jsx
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 import BpmnModeler from 'bpmn-js/lib/Modeler';
-import { emptyBpmn } from '../../assets/empty.bpmn';
+import { emptyBpmn } from '../../../assets/empty.bpmn';
 import propertiesProviderModule from 'bpmn-js-properties-panel/lib/provider/camunda';
 import camundaModdleDescriptor from 'camunda-bpmn-moddle/resources/camunda';
-import PropertiesView from '../properties-panel/propertiesView/PropertiesView';
-import arkRPA_ModdleDescriptor from '../../assets/modelerPropertiesExtensionRPA/ark-rpa';
-import parser from '../../api/parser.js';
+import PropertiesView from '../PropertiesPanel/PropertiesView/PropertiesView';
+import arkRPA_ModdleDescriptor from '../../../assets/modelerPropertiesExtensionRPA/ark-rpa';
+import parser from '../../../utils/parser.js';
 import convert from 'xml-js';
 
 import './BpmnModeler.css';
