+++ client/src/components/content/PropertiesPanel/PropertiesPanelApplicationDropdown/PropertiesPanelApplicationDropdown.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Select } from 'antd';
-import '../propertiesView/PropertiesView.css'
+import '../PropertiesView/PropertiesView.css';
 
 const { Option } = Select;
 
@@ -8,7 +8,7 @@ const { Option } = Select;
  * @class
  * @component
  * @classdesc Renders the application-dropdown based on passed list of applications.
- * @example 
+ * @example
  * let applicationList = ['MS Excel', 'Browser'];
  * let handleApplicationSelection = (event) => return 'successfully handled application selection';
  * return (
@@ -20,17 +20,19 @@ const { Option } = Select;
  */
 
 export default function PropertiesPanelApplicationDropdown(props) {
-    return <>
-        <Select
-            showSearch
-            style={{ width: '100%', marginTop: '10px' }}
-            placeholder="Please select application"
-            onChange={props.onApplicationSelection}
-            defaultValue={props.currentSelection}
-        >
-            {props.applications.map((application) => (
-                <Option value={application}>{application}</Option>
-            ))}
-        </Select>
+  return (
+    <>
+      <Select
+        showSearch
+        style={{ width: '100%', marginTop: '10px' }}
+        placeholder='Please select application'
+        onChange={props.onApplicationSelection}
+        defaultValue={props.currentSelection}
+      >
+        {props.applications.map((application) => (
+          <Option value={application}>{application}</Option>
+        ))}
+      </Select>
     </>
+  );
 }
