+++ client/src/components/content/PropertiesPanel/PropertiesView/PropertiesView.css
