+++ client/src/components/content/PropertiesPanel/PropertiesView/PropertiesView.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import PropertiesPanelBuilder from '../propertiesPanelBuilder/PropertiesPanelBuilder'
+import PropertiesPanelBuilder from '../PropertiesPanelBuilder/PropertiesPanelBuilder';
 import './PropertiesView.css';
 
 import { Typography } from 'antd';
@@ -11,8 +11,8 @@ const { Title } = Typography;
  * @component
  * @classdesc This class decides which sidebar is displayed. It updates itself depending on the number of selected BPMN elements.
  * @example return (<PropertiesView />)
- * 
- * @description Initializes state based on properties. 
+ *
+ * @description Initializes state based on properties.
  */
 
 export default class PropertiesView extends Component {
@@ -24,7 +24,7 @@ export default class PropertiesView extends Component {
     };
   }
 
-  /** 
+  /**
    * @description
    * On a changed selection, the selected element changed in the components state.
    * We also update panel via .setState, if currently selected element changed.
@@ -60,20 +60,24 @@ export default class PropertiesView extends Component {
     const { selectedElements, element } = this.state;
 
     return (
-      <div class="sidebarWrapper">
+      <div class='sidebarWrapper'>
         {selectedElements.length === 1 && (
           <PropertiesPanelBuilder modeler={modeler} element={element} />
         )}
 
         {selectedElements.length === 0 && (
           <span>
-            <Title class="label-on-dark-background">Please select an element.</Title>
+            <Title class='label-on-dark-background'>
+              Please select an element.
+            </Title>
           </span>
         )}
 
         {selectedElements.length > 1 && (
           <span>
-            <Title class="label-on-dark-background">Please select a single element.</Title>
+            <Title class='label-on-dark-background'>
+              Please select a single element.
+            </Title>
           </span>
         )}
       </div>
