+++ server/server.js
@@ -31,9 +31,9 @@ if (!isDev && cluster.isMaster) {
   app.use('/rpa-framework', rpaFrameworkRouter);
 
   // All remaining requests return the React app, so it can handle routing.
-  app.get('*', function (request, response) {
+  /* app.get('*', function (request, response) {
     response.sendFile(path.resolve(__dirname, '../client/build', 'index.html'));
-  });
+  }); */
 
   app.listen(PORT, function () {
     console.error(
